Changes after review with @bagder:
- int in data cfilters control/query methods as arg
- shift HAProxy setup filter "below" SSL
- fail HAPRoxy if SSL is in place
- remove struct http_conn
- quiche remove version field
- quiche. drain easy handles need expire, look at http2 code

+++ lib/cf-socket.c
@@ -1188,7 +1188,7 @@ static void cf_socket_active(struct Curl_cfilter *cf, struct Curl_easy *data)
 
 static CURLcode cf_socket_cntrl(struct Curl_cfilter *cf,
                                 struct Curl_easy *data,
-                                int event, long arg1, void *arg2)
+                                int event, int arg1, void *arg2)
 {
   (void)arg1;
   (void)arg2;
